Fix #1839 Sunrise GFX9000 emulation does not emulate IEHM flag
https://github.com/openMSX/openMSX/issues/1839

Strictly speaking IEHM was emulation was attempted, but it had an
off-by-one bug. Instead of calculating the time of (start of the) the
_next_ line it calculated the start of the _current_ line. (which then
later got wrapped to the next frame, see patch for details).

At the time we did not have any test case, so it's not unsurprising that
the implementation contained a bug. So thanks a lot to "cave-fish" for
creating a test case! With this test case it was fairly easy to find and
correct the bug.