Fix #1840 Sunrise GFX9000 emulation schedules LINE command completion incorrectly
https://github.com/openMSX/openMSX/issues/1840

It was only wrong when using an IRQ on command completion, not when
polling the status register.

Caused by an off-by-one error. E.g. when setting NX=10, the command
draws a line containing 11 (=10+1) pixels, not 10. But internally we
scheduled a sync-point only 10 units ahead in the future (not 11),
thus when this sync-point triggers and we execute the (partial) LINE
command, there's still one pixel remaining and thus we did not yet
trigger the IRQ.

Thanks a lot to cave-fish for reporting this problem and especially for
creating a test-case (including source code)!!