[1/15] remove const_cast from DeviceConfig.cc
(Many) extension XML config files have "any" for the value of the
primary and/or secondary slot attribute. Once we insert such an
extension into a concrete slot in a machine, we replace those
"any"-values with the actual slot number (e.g. this is needed when we
later restore this setup from a savestate). Another example is when a
rom lists several possibilities for sha1sums, after loading a concrete
rom we record the actual sha1sum.

Before this patch, the code that made these modifications in the
configuraton only had access to a **const** `DeviceConfig`. And then it
had to cast away this const. That's not nice. So this patch instead
passes non-const objects in all required places. In many cases the
caller (usually) doesn't modify the object. But at least now we no
longer hide the fact that in some cases we DO make modifications.