[3/15] remove const_cast from MSXMultiXxxDevice
By far the majority of the MSXDevices has a fixed name. Only
MSXMultiIODevice (when you have multiple devices using the same IO port)
and MSXMultiMemDevice (when you have multiple devices present in the
same 16kB page) have a dynamically changing name.

We can query the name via this method:
    virtual const std::string& MSXDevice::getName() const
There are lots of possible trade-offs, but we've chosen to:
* make the method virtual, so that we can customize the implementation
  for the MSXMultiXxxDevice classes (and lazily calculate the name).
* Return a `const string&` instead of `string` by value. Because by far
  in the majority of the cases we avoid making a copy of the string.
* In MSXMultiXxxDevice we reuse the storage of the 'deviceName' variable
  in the base class for the (in this class) dynamic name.

Because of the combination of these choices we had to use a const_cast
in MSXMultiIODevice and MSXMultiMemDevice. In this patch we avoid that
const_cast by making `deviceName` mutable. Thus indicating that
`getName()` is indeed "logically" const, but `deviceName` is
"physically" allowed to change (acting like a lazily calculated cache).