[4/15] remove const_cast from InputEventGenerator
In joystick SDL events we translate the joystick instance IDs (unique
numbers for the whole session) to "logical" IDs (small numbers, reused
when unplugging / re-plugging a host Joystick). We do this translation
in the SDL event struct itself (that required no or minimal changes to
the openMSX code using those Joystick events).

Initially we didn't have this translation and then it was fine to pass
around **const** SDL_Events. When we added this translation we (had to)
cast away const. But it's cleaner to admit that the objects are no
longer not being modified, so pass them around as non-const.