[5/15] getNbSectors() and readSectors() are non-const ...
At first sight it makes sense that the method
   size_t SectorAccessibleDisk::getNbSectors() const
is const. Querying the number of sectors in a disk should not change
anything about that disk.

However for disk images like `DMKDiskImage` the number of sectors is not
clearly defined (some tracks have fewer or more sectors, invalid
sectors, ...). And for some calculations we need a fixed number of
sectors per track (e.g. to translate a "logical" sector number into
cylinder/head/sector). So we use a combination of heuristics
(implemented in detectGeometry), but some of these need to make
modifications to the state. Mostly caches, so in principle we could make
that state "mutable". But in this case it was easier to just make
`getNbSectors()` non-const (most of the callers anyway already had a
non-const object).