[6/5] avoid const_cast in ImGuiPalette
The 'fixed' palette is stored in a global constant. When this palette is
selected we have checks in place to ensure we can't edit this palette.
Though the type system doesn't see those checks, so we had to use a
const_cast.

That's not pretty, so in this patch we instead make a copy. (On modern
x86 copying 32 bytes only takes 1 or 2 load/store instructions, so the
cost is negligible).