[7/15] cleanup initializing 'ram' in 'AmdFlash'
The `SRAM` content can be freely read, I mean we can directly read from
the internal memory buffer in the SRAM object (via a const operator[]).

But when we write to the SRAM, we want those change to be also written
to disk. So in the `write()` method we also schedule a write to disk (we
don't write to disk each time, instead we try to do a single disk write
for a large group of write() calls).

There is one exception. Internally in the `AmdFlash` class we also use
an `SRAM` object to store the flash content. If the flash content gets
modified we also want to write that to disk. But we avoid writing an
unmodified flash to disk (with initial content from a rom file). So we
cannot initialize the SRAM with the normal write() calls (those _would_
schedule a disk write).

Before this patch we used the `const byte&` returned from `operator[]`
and then cast away the const. That's an ugly hack that does allow to
modify the content without scheduling a disk write. This patch adds a
new method `getWriteBackdoor()` with the same semantics, but with a more
explicit name, so that it's clear something 'special' is going on.