[8/15] remove const_cast from Rom, RomFactory
The 1st patch in this series already talked about how we record the
actual sha1sum for a rom file (e.g. this may not be fixed because the
config specifies multiple possibilities). The same is true for the
actual mapper type: the actual mapper type may change because of patch
files (and then the guess-mapper-type function gives a different result)
or the user may have explicitly specified the type on the command line.

This patch passes more objects around as non-const to be able to remove
more const-casts (because we need to modify the config for the above
features).