[9/15] remove const_cast from RomDatabase parser
The result of parsing the rom database (softwaredb.xml) is a data
structure containing pointers into the original file. These point to
strings, but the strings are (in-place) modified with zero-terminators.

Our XML-parser is event driven. You get callbacks for opening/closing
tags, for attributes, for text, ... These callbacks receive string_view
parameters (pointing to the original file content). Before this patch we
would cast away const (string_view offers a read-only view only) and
then add the zero-terminator.

Though our XML parser already had an **option** to add zero terminators.
We just didn't enable that option. And then we don't have to do it
ourselves anymore. I think one reason why we didn't use that option is
that it adds more zero-terminators than we need. E.g. it also
zero-terminates tag-names and attribute values/names. While in this
specific case we only need it on (most) text fields.

So this patch removes our zero-termination and replaces it with enabling
the option in the parser. That does a little bit more work, but the
performance difference is minimal (I couldn't measure the difference).