[10/15] remove const_cast from XmlInputArchive
When loading a savestate (.oms files are compressed xml files) we allow
the tags to be used out-of-order (within a single tag). To keep track of
which tags have already been used we modify the input (change the name
of the tag).

Most of the load code is using 'XmlElement' in a read-only way. That's
why it was using **const** objects everywhere. Except for the "marking"
feature discussed above. Before this patch we worked around that via a
const_cast. This patch removes just enough of the 'const' qualifier so
that this cast is no longer needed.