[15/15] remove const_cast from V9990PxConverter
In the V9990 P1 and P2 modes, in some cases, palette index 0 is replaced
by the backdrop color. To render this efficiently we modify index 0 in
the palette to instead contain that backdrop color.

Logically rendering should not modify the palette. That's why the
renderer classes received **const** versions of the palette. But to
perform the above rendering optimization we do *temporarily* modify the
palette (and then restore it). This was done via a const_cast.

This patch instead passes a non-const palette and removes the
const_cast.