Cleanup after eeec2828a4a
In commit eeec2828a4a we passed `DeviceConfig` by value to the `Rom`
constructor. That was done to be able to pass a temporary (that's not
possible with a non-const reference parameter). But looking back that
was a bit weird. So in this patch we instead _do_ pass via non-const
reference, and solve the 'temporary variable' problem in a different
way.

Because we no longer make a copy of DeviceConfig, more parameters needs
to be adjusted for const vs non-const. But this was the intention from
the start (it makes it better visible which configs get potentially
modified).