[7/7] Replace putenv() with setenv()
The former takes a non-const `char*` parameter (and we pass a
string-literal, thus we need a const-cast). The latter takes 2
`const char*` parameters.

The downside is that setenv() does not exist on Windows. Though luckily
we already implemented a setenv() for Windows (by using _putenv_s()) in
some other location (main.cc). This patch makes that helper function
reusable from everywhere.