Implemented export option for 'DebuggableEditor'
Right-click then select 'Export ...'. This opens a new window with
several options:
* Select the range to output (all, or [begin, end]).
* Raw or formatted output, lots of options for formatting. ("Lots" but
  not "every" possible formatting).
* Output to clipboard or to file.

I added lots of help-markers on the different options.

This should implement the 2nd part of:
    #1848 Missing Sprite Data textbox on the Sprite Viewer widget
    https://github.com/openMSX/openMSX/issues/1848
Instead of implementing something specific for the sprite (or tile)
viewer, this export option works on all debuggables (with more flexible
output options). But sure, you do need to type in the address.