Apply clang-tidy readability-redundant-inline-specifier
Methods implemented in the class definition itself are implicitly
'inline'. In the past the compiler would also interpret 'inline' as a
hint to perform function inlining. But nowadays the meaning is mostly
only to indicate weak-linking. (It's still a minor hint to perform
function inlining, but that hint is getting weaker with newer compiler
versions, in general compilers are very good at making inline decisions
by themselves, we don't need to write manual hints anymore).

So this patch removes all manual redundant "inline" annotations.