[1/4] replace nibble/byte/word with uint{4,8,16}_t
The meaning of 'word' is highly context dependent. In MSX context it
often means 16 bit, on modern systems it's often 32 or even 64 bit. If
we instead use 'uint16_t' the meaning is unambiguous. And it also states
we want to interpret it as an unsigned integer.

Almost always 'byte' means 8-bit (there are rare exceptions). Though we
still like to replace 'byte' with 'uint8_t'. As above that makes it
clear we interpret it as an unsigned integer. And this also avoids
confusion with the c++17 'std::byte' type:
    https://en.cppreference.com/w/cpp/types/byte
The latter is more like a storage-type, a container of (usually) 8 bits,
but without support for arithmetic operations. And we do want 8-bit
unsigned arithmetic, so for us 'uint8_t' is a better fit.

The meaning of 'nibble', half a byte, 4 bits, is usually clear. Though
for consistency with the above, we'll also rename it to 'uint4_t'.

--

The openMSX source code heavily uses 'byte', a little bit less 'word',
and occasionally 'nibble'. In the end I'd like to replace all of these.
This patch series makes a start but doesn't fully achieve that goal yet.

This first patch does already replace all uses of 'nibble' with
'uint4_t' (and it also replaces 'byte'/'word' in the files that it
touches).