Fix unittest after 76d76e22b3
The output for strCat(my_float) has changed after the above commit. That
is because the statements
    std::to_string(my_float)
and
    my_string_stream << my_foat;
both produce valid string representations for the given float, but they
are not identical. For more detail, see the example on:
    https://en.cppreference.com/w/cpp/string/basic_string/to_string

I don't believe we currently rely on this behavior. I think we just
added strCat(float) for completeness. And both behaviors are fine.

In the future, when we want more control over the exact formatting, we
can use std::format() (part of c++20, support in gcc-13).