Resolve -Wodr warning in LTO build
Both 'OSDWidget.cc' and 'ImGuiSettings.cc' defined a _different_ struct
'Rectangle' with the _same_ name. That's a One Definition Rule (ODR)
violation.

Fixed by placing (both) in an anonymous namespace. This gives the class
internal linkage, meaning the name is private in the .cc file it's
defined in. Thus avoiding the ODR violation.

This fixes:
    One-Definition-Rule for link time optimization is broken by struct Rectangle
    https://github.com/openMSX/openMSX/issues/1873