Fix formatting of 'speed' in GUI status bar
Commit 76d76e22b3fa699440565764da80337fb051a075 changes how floating
point numbers are formatted via strCat(). I thought we didn't use that,
but apparently we do.

The speed value is formatted via:
    strCat(std::round(value), '%')
round() is expected to return a whole number, nevertheless the type of
that number is still a floating point. After the above commit, float
values are always formatted with 6 decimal places. Before it depended on
the actual value.

One fix could be to restore the behavior of strCat(). But for this
specific case it's better to not print the (whole) value as a floating
point but as an integer. So this patch changes std::round() to
std::lrint(). That fixes the formatting and it's also faster.