FrameBufferObject: generalize usage pattern
Before this patch our FrameBufferObject class offered a push() and pop()
method. It could only be strictly used as a stack: you can activate a
FBO (via push()), and later upon deactivation (via pop()) you always
restore the FBO that was active before the push() (if any).

This is a good fit for the current needs. Though soon we'll add a usage
pattern that doesn't follow this stack behavior. So this patch replaces
push() and pop() with getCurrent() and activate(). Building the stack
behavior should then be done manually.