Scalers: simplify region creation
A typical scale-algorithm deals with a fixed scale-ratio. For example it
scales an MSX image of a certain width (e.g. screen 5 vs screen 7) and a
certain height (interlaced or not) to a host output window of a certain
size. So it may use a different approach for screen 5 vs screen 7 and
interlaced or not.

Though because of screen-splits it can happen that within one frame
there are different regions with different widths. And in fact we model
the vertical borders as regions with width=1. So a mix of different line
widths is actually a very common situation. Because of this the frame is
split in regions with equal input-width.

Let's now look at the ratio between input and output height. Suppose the
MSX screen is (de-)interlaced. That means we (currently) emulate it as a
frame with 480 lines. And suppose we've selected scale-factor=3, this
corresponds to a host window with 720 lines. In other words for every 2
MSX lines we need to produce 3 host lines.

In the past we (also) had the scalers implemented in software (I mean
non-openGL). For those it was important to really handle the above
scenario in groups. So input lines in groups of 2, to produce output
lines in groups of 3.

Now we only have the openGL scalers. These shaders always run per output
pixel (and thus also per line). So in the above scenario there's no need
anymore to group the output lines per 3.

And this allows to simplify the region-creation algorithm. Before we
took the ratio between input and output lines into account. But that's
no longer needed. The new algorithm only groups regions on equal
input-line-width.

--

This patch also simplifies and generalizes the parameters we pass to the
GLScaler::scaleImage() method.
* A region starts and ends at a certain line. Before we passed these
  lines both in the input and output coordinate system. But that's
  redundant, now we only pass it as input-coordinates.
* We passed the input-width and input-height as 2 separate parameters,
  in this patch we group these into a single `ivec2 srcSize` parameter.
* We passed only the output width, but in this patch we generalize that
  to the full output size `ivec2 dstSize`. This is a preparation for
  later patches, where we'll add non-integer (and non-fixed
  aspect-ratio) scaling.