Generalize setting up MVP matrix in scaler
For the scalers we setup an orthographic projection as the
Model-View-Project (MVP) matrix. This matrix should map coordinate (0,0)
to the top-left host pixel and (width-1, height-1) to the bottom-right
pixel. The same matrix is used for several other purposes besides
scalers. So we calculate it once and share it. This is the
`glContext::pixelMvp` array.

In later patches in can be that a scalar runs in two steps. For example
the HQ scaler can only deal with integer scale factors. We'll do
non-integer scaling by:
* First upscale the MSX 320x240 by a factor 4x to 1280x960. (Choose the
  factor 2x, 3x or 4x to arrive close to the desired output size).
* Then scale this 1280x960 image to the actual window size (e.g.
  1440x1080).
But that means the transformation matrix we need in the first scaling step
is different from the `glContext::pixelMvp` matrix (because that one
depends on the actual window size).

In this patch we prepare for this future possibility by always
recalculating the correct transformation matrix. This is anyway a
reasonably cheap calculation.