Rewrite GLDefaultScaler (=fallback scaler)
In some cases the selected scale-algorithm is not used(*1). For example:
* simple scaler with blur=0 and scanline=0
* HQ scaler with ratio < 2 (e.g. interlace and scale_factor=2).
* ...
In those cases we instead use the "default" scaler as a fallback.
(*1) Maybe in the future we can reduce the number of cases where we use
     the fallback

Before this patch that default scaler would just do plain-texture
mapping, even with nearest-neighbour interpolation. For integer
scale-factors that's just fine. But we want to extend scaling to
non-integer scale-factors. And then this default scaler looks very ugly.

This patch replaces the default scaler with a more sophisticated one. It
does resampling weighted by the covered area of input vs output pixels.
I'll try to explain.
* Suppose the input has A x B number of pixels.
* The output has C x D number of pixels.
* There are no restrictions:
  * The input can be larger or smaller than the output.
  * The aspect ratio of input and output can be different.
  * So e.g. it could be that the input-width is smaller than the
    output-width, but the input-heigh is larger than the output-height.
* Now imagine a square and divide that into A x B small rectangles.
  These correspond to the input pixels.
* Divide that same square also into C x D small rectangles. These are
  the output pixels.
* Every rectangle corresponding to an output pixel covers one or more
  rectangles corresponding to input pixels.
* Some input pixels may be fully covered, some only partially covered,
  or by different amounts.
* We now calculate the color of the output pixel as the average of all
  covered input pixels, weighted by the area of the coverage.

This scaler has some nice properties:
* When downscaling (many input pixels mapped onto fewer output pixels)
  we take the color of all input pixels into account.
  * Nearest neighbour interpolation would skip over pixels.
  * Bilinear interpolation with mip-mapping would take all pixels into
    account, but requires extra steps to calculate the different
    mip-maps.
* When upscaling (by a large amount), pixels remain "square". Only at
  the boundaries (corresponding to different input pixels) there's a 1
  pixels wide interpolation.
  * Bilinear interpolation would look very blurry,

At the moment there are still only integer scale factors. So the new
scaler doesn't have an advantage yet. But we'll soon enable non-integer
factors.

Note that a very similar scale algorithm will be used later to turn
integer-only scale algoithms into arbitrary factor scalers (the 2nd step
that's described in the previous commit message).