Extend scalers to non-integer scaling
Added a new method:
    virtual gl::ivec2 GLScaler::getOutputScaleSize(gl::ivec2 dstScreenSize);
Given an desired output size, each scaler can return what output size it
can actually produce:
* Scalers that can do arbitrary scaling will return the size unchanged.
  For example the TV and simple scaler do this.
* Scalers that work best for one specific output size, return that size.
  For example the RGB scaler works best with an output of 960x720.
* Some scalers can only do a fixed set of sizes. For example the HQ
  scaler can produce 640x480, 960x720 or 1280x960.
* Some scalers can only do integer multiples of 320x240.

After calling getOutputScaleSize(), we run the scaler to produce that
output-size.
* If this was also the desired size, then we're done.
* If not then we rescale that intermediate result to the desired size.

That 2nd step is done with a similar weighted area scaler as described
in the previous patch.

---

This patch prepares the scalers for arbitrary output size. But because
the scale_factor settings can only have 3 fixed values (2, 3 or 4) we
can't really use this yet(*1). The next patch will allow to freely
resize the window.
(*1) Maybe only the RGBTriplet scaler is improved already because it
     now always produces an intermediate result at 960x720.