Fix use-after-free bug in TclParser
This fixes:
  #1878 Console crashes openMSX with SIGABRT
  https://github.com/openMSX/openMSX/issues/1878

TL;DR
  Remove the member variables `parseStr` and `offset`, and instead pass
  them as parameters to various functions.

`parseStr` and `offset` are needed in several different member
functions. That's why (before this patch) we made them member variables.

The `parse()` function can get called recursively. This is done while
parsing a command that contains subcommands. On such a recursive call
we're processing a different `parseStr` and at a different `offset`.
After that recursive call returns we need to go back to the previous
`parseStr` and `offset`. Before this patch this was done via the
`ScopedAssign` utility.

Conceptually the above works fine:
* We have a compact syntax (no need to pass `parseStr` and `offset`) to
  the different functions.
* Using `ScopedAssign` ensures the value of these variables follows the
  semantics of a recursive function call: we change the value when
  entering the function and restore the original value when leaving.

However, `parseStr` is a `std::string`, and we also create
`std::string_view` objects to substrings of `parseStr`. Usage of
`ScopedAssign` ensures that the value of `parseStr` is correct. But it
does NOT guarantee that after calling and returning from `parse()` the
*address* of the string remains unchanged. Thus `string_view` objects
become invalid after a call to `parse()`.

So in this patch we go for the more verbose syntax:
* Remove the `parseStr` and `offset` member variables.
* And instead pass them as parameters to all functions that need them.