Made FileOperations::getFilename() safer to use with zero-terminated strings
Before this patch `getFilename()` took a `std::string_view` as input
parameter and also returns a `std::string_view`. It always returns the
tail of the input string (can also be the full input). So if the input
is zero-terminated, the result will also be. Though this isn't clear at
all from the type-system. The result is a `std::string_view` and in
general that type is not guaranteed to be zero-terminated.

In this patch we add an extra overload:
  zstring_view getFilename(zstring_view);
So if the input is guaranteed to be zero-terminated, we now retain that
guarantee for the result.

And to solve ambiguous overload compilation errors when called with a
`std::string` parameter we add a 3rd overload that takes a
`std::string`. That one delegates to the `zstring_view` overload, so it
also retains the zero-termination guarantee.