[1/2] Change TclObject list size/index from 'unsigned' to 'size_t'
In Tcl version 8, the size of a list is represented by the 'int' type
(32-bit). Therefor in our TclObject wrapper class we used 'unsigned'.
But in Tcl version 9 this type was changed the 'Tcl_Size' which is a
typedef for 'ptrdiff_t' (64-bit). So that would cause truncation in our
TclObject wrapper.

As a preparation for that (eventual) change, this patch changes our type
from 'unsigned' (32-bit) to 'size_t' (64-bit). That's OK when using Tcl8
and (later) required when using Tcl9.

I checked all users of TclObject that do something with list sizes or
indices, and adjusted them all. This didn't need to many changes (much
less than I was expecting), also because many users already used 'auto'
to make the code type-agnostic.