Preparation for Tcl9: use TCL_CHANNEL_VERSION_5
We set a custom Tcl_Channel to redirect the output from Tcl to the
openMSX console. The simplest channel is sufficient (just capture the
output), so untill now we used the first supported channel version.

Tcl9 dropped support for channel version smaller than 5. Tcl8.6 does
already support that version. So if we use that in openMSX, we can
support both Tcl8 and Tcl9 with the same source code.

See also the discussion here:
    https://github.com/openMSX/openMSX/pull/1905