Be a bit more robust about machine configs changing.
There was a use case of someone trying to edit his custom machine config whilst
openMSX was running. One of the things he did was rename the file and then he
tried to load the renamed config. But it also meant the config file of the
currently running machine was gone. And our code had the assumption this would
always still be available. So I tried to improve that a bit and showed plainly
the config name if the config file of the currently running machine is not
found.
I also added an extra cache wipe and rescan of the folder in case the info
isn't found.

This should of course be a rare case.