Fix crash on debug-write to TrackedRam
Reported by pvmm:
    https://github.com/openMSX/openMSX/issues/1914

For the reverse susbsystem, once per second, we take a snapshot of the
full MSX machine, including the full RAM content. Some RAM blocks are
large and don't change very often (e.g. the OPL4 sample ram, or the
V9990 video ram (when not running V9990 software)).

To speedup taking snapshots of such RAM blocks we keep a bool variable
that indicates there has been at least one write since the last
snapshot. This variable is set on _all_ writes (so even if we set a
different value and then restore the previous one).

But there was a bug: debug writes did not set this bool. And then when
openMSX is build in debug mode we have an (expensive) check that
verifies when that bool variable is false (meaning there was write
recorded) the content of the block is indeed the same as the content of
the block in the previous snapshot. And that check found the bug.