Implementation of saving media in setups.
Code written by Wouter. Just committing here so I can build on top of it.

Here follows some more guiding text for this commit, written by Wouter.

The method getMediaInfo() already existed, setMedia() is new.
The idea is that the result of getMediaInfo() contains (more than) enough
information to insert the same media in another machine.
By far the largest part of the patch is to implement setMedia() for all media
(disk, cartridge, cassette, hd, cdrom, scsils120, laserdisc). The patch also
does some renames.

From a high level each setMedia() does something like this:
* parse the "info" parameter (this is a Tcl dict), often this is just: look at
  the value of the "target" entry.
* if there is no media inserted (often, but not always, this corresponds to an
  empty string for "target") then do the equivalent of "removeMedia()"
* otherwise do the equivalent of "insertMedia()" (possibly with extra options
  like -romtype and -ips patches)

Many devices already had a "removeMedia()" and "insertMedia()" (typically with
a different name). And then the implementation of the new setMedia() method is
only about 5 lines of code. With ROMs there was no "removeMedia() /
insertMedia()" so part of this patch will first factor that out.

In principle, if you only look at the "restore-setup" use case, then it is
redundant that setMedia() can also do a "removeMedia()" (because we always
start from an empty machine). But who knows, maybe in the future we can also
use this to simplify the GUI code a bit (a more generic way to handle all
media... now there is a big block of code in the GUI for each media type that
is very similar but still not enough to be able to factor out).

For a cartridge slot, getMediaInfo() can return information if an extension is
inserted. But setMedia() cannot duplicate that situation. I mean getMediaInfo()
does show information about extensions, but setMedia() cannot insert
extensions.

And I did this deliberately: we don't need it today (extensions have already
been handled in an earlier level) and doing it anyway could possibly cause
problems: I'm thinking of the following scenario:
* External disk drive inserted in carta (this then gives diska as media)
* A disk image inserted in that diska.
* With a "restore setup" we will first restore the extensions (so carta)
* Then the media in a RANDOM order (I don't mean deliberately random, but you
  have no guarantee that either diska or carta will happen first)

Suppose we do "diska" first: so far so good, then we restore the inserted disk.
And only then media "carta". And now suppose that setMedia() would also restore
extensions.
* Because there is already an extension in carta (done by level 2) we will
  first remove it.
* And immediately afterwards we will re-insert the same extension.
I don't know what that does to the media in diska. My guess is that the disk
will be lost.

So it's a shame that setMedia() is not completely generic, but today this seems
to me the most pragmatic solution.
getMediaInfo() also returns information about the cassette player state
(stopped/playing/recording) and they don't want to replicate either.