Fix TTFFont::getSize()
When porting from SDL2_ttf to SDL3_ttf, the function `TTF_SizeUTF8()` was
replaced with `TTF_GetStringSize()`. The former returns `0` on success, `-1`
on error. The latter returns `true` on success, false on error. So when
switching between these two, we have to negate the condition in the
surrounding `if`.

Usually this code didn't trigger, except when printing multi-line
strings via "osd create text".