Allow user to set default setup in the GUI.
- when saving a setup, the user can tell openMSX to use the saved setup as
  default
- when loading/running a setup, the user can specify to run it also when
  starting up
- mark the setup entry with [default] if it's the default setup

For the latter, I had to extend FileListWidget to have a hook to override the
display name function. I only changed that much to make it sufficient for this
new case. As the display name is now dynamic, everywhere where displayName was
used the call to the display function must be executed. And when we just want
the default display name (fileName without extension, in almost all cases), we
call the default display name function explicitly.