Only send QuitEvent once
SDL behave slightly different in KDE, GNOME, and possibly also other WMs:
* KDE: sends both SDL_QUIT and SDL_WINDOWEVENT_CLOSE (for the main window)
* GNOME: only sends SDL_WINDOWEVENT_CLOSE
* Possibly in other cases only SDL_QUIT is send?

In openMSX we listen to both events, translate them to a single
`QuitEvent`, and different parts of openMSX can then respond to that.
Some examples are:
* The Tcl 'after quit' callback
* (Soon) the save-setup-on-exit functionality.

The trigger for this patch was that we saw the latter save the setup
twice (at least when using KDE).