Fix 'osd info -mousecoord' when the main-window doesn't have focus
The OSD stuff was implemented when openMSX only had a single window
(before we added Dear ImGui with multiple viewports). At that time 'osd
info -mousecoord' worked fine.

Internally 'osd info -mousecoord' queries the coordinates via
SDL_GetMouseState(). That function returns the coordinates relative to
the *focus* window. So if there's only one window that's fine. But if
there can be multiple windows we get issues like described here:

    https://github.com/openMSX/openMSX/issues/1920

This patch changes `gl::ivec2 SDLVideoSystem::getMouseCoord()` to
instead return an `std::optional`. When the main window doesn't have
focus it now returns `std::nullopt`.

`osd info -mousecoord` uses that function internally. Now when it gets
`nullopt` it will return the coordinates {inf, inf} to the calling Tcl
script. This is the same convention that we already use for when the
cursor is invisible (e.g. via `pointer_hide_delay`). That's for sure
outside the queried OSD widget.

The request in issue#1920 was to (also) not send mouse events (to Tcl)
when clicking outside the main window. But that's not what this patch
implements. This patch makes another (also necessary) fix, and which may
already be sufficient.