Make ellipsis usage in menus consistent and to guidelines.
We use the ellipsis without space in front, to indicate that selecting this
menu item will not directly trigger an action, but more details will have to be
provided in a new window. So, the main purpose is to indicate that the user
shouldn't be afraid selecting the options as no direct action will take place.

See also https://learn.microsoft.com/en-us/windows/win32/uxguide/cmd-menus#using-ellipses