macOS: Fix CI build for pull requests.
Code sign identity is now specified in a GitHub Actions repository variable.
This will allow forks to easily do their own code signing. If not configured,
then the code is not signed but the build will succeed.

See 244eabd for the previous fix, which was lost in 2d060e7 / a281e5d.