txt2omr: Ignore non-decimal id values
In older OMR files there were only decimal ids, but it seems the OMR
format got changed in an incompatible manner. This script needs to
create unique new decimal ids, so there won't be a conflict with
non-decimal ids anyway and it's safe to ignore them.

This addresses issue #1932, but I don't want to close that issue here,
since it might require further discussion.