[1/2] Simplify passing EmuDuration/EmuTime
We used to write
    void myFunction(EmuTime::param time);
this patch simplifies that to
    void myFunction(EmuTime time);

Internally the 'EmuTime' type stores a single 64-bit value.

'EmuTime::param' is a typedef that expands to:
* Just 'EmuTime' on 64-bit systems, so pass EmuTime directly by value.
* 'const EmuTime&' on 32-bit systems, so effectively pass a pointer.
This trick was used so that on both systems passing an EmuTime parameter
only takes a single register.

On 64-bit system we can then access the time directly. On 32-bit systems
we need (one or two) extra instruction(s) to dereference the pointer.
Though most functions that take an EmuTime parameter only forward that
parameter to deeper functions, and then this is a good trade-off (pass a
single register vs extra indirection).

And exactly the same story for EmuDuration.

--

In this patch series we simplify this and just always pass 'EmuTime' by
value:
* This simplifies the notation in literally thousands of places.
* This is exactly the same as the current behavior on 64-bit systems.
* On 32-bit systems, depending on the exact context, this can be
  slightly less efficient. But 32-bit systems are rare nowadays.
* In exceptional situations the behavior between both approaches can be
  different (when the underlying EmuTime object that we refer to
  changes). And this can lead to subtle bugs. (For a long time already)
  we don't test anymore on 32-bit system. So these bugs probably won't
  get discovered. Better to eliminate this possible source of bugs.

But the main motivation is to simplify the code.