macOS: Update Xcode to 16.4.
Xcode 16.4 requires macOS 15. The device support for macOS 15 is the same as
macOS 14, except for Macbook Air 2018 and Macbook Air 2019.

So this should not significantly affect support for users wishing to compile
openMSX on old Intel Macs. This change does not affect the minimum requirement
for running openMSX binaries, which remains 10.15.

New compiler features since Xcode 15.4:
https://developer.apple.com/documentation/xcode-release-notes/xcode-16-release-notes#Apple-Clang-Compiler
https://developer.apple.com/documentation/xcode-release-notes/xcode-16_3-release-notes#Apple-Clang-Compiler
https://developer.apple.com/documentation/xcode-release-notes/xcode-16_4-release-notes#Apple-Clang-Compiler