Don't loose knowledge about length of string-literal
Consider this code snippet:
    void f(std::string_view s);
    ...
    f("abc");                          // 1
    f(myCondition ? "abc" : "12345");  // 2

The function f() takes a `std::string_view` parameter. In the first call
to f() we pass a string-literal "abc", this has type `const char[4]`.
And gcc/clang can convert this literal to a string_view without calling
strlen(). Visual studio does not seem to be good at avoiding strlen().

In the second call to f() we choose between "abc" and "12345". These
respectively have types `const char[4]` and `const char[6]`. But after
the ternary we need a common type and that is `const char*`. In other
words: we've lost the information on the length of the buffer containing
the literal. And (in specific tests) I see that gcc now does call
strlen() (clang could still avoid it).

In this patch we rewrite the above pattern to the following (add the
"sv" suffix):
    f(myCondition ? "abc"sv : "12345"sv);  // 3

Now both string-literals (with different length) are individually
converted to string_view first before the ternary. And that retains the
length information.

I did verify that in this case both gcc and clang (and sometimes also
msvc) are able to avoid calling strlen().